Fix warnings.

+++ tests/crypto_rng.c
@@ -1,3 +1,24 @@
+/*
+ * Copyright (C) 2008 Free Software Foundation
+ *
+ * Author: Nikos Mavrogiannopoulos
+ *
+ * This file is part of GNUTLS.
+ *
+ * GNUTLS is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * GNUTLS is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with GNUTLS.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
 
 #ifdef HAVE_CONFIG_H
 # include <config.h>
@@ -26,7 +47,7 @@ rng_init (void **ctx)
 }
 
 int
-rng_rnd (void *ctx, int level, void *data, int datasize)
+rng_rnd (void *ctx, int level, void *data, size_t datasize)
 {
   memset (data, 1, datasize);
   return 0;
